Optional ContentType header for AddDocument
I was encountering an issue when trying to upload certain PDFs.  I was able to upload them via the AddDocument method, but when I tried creating the Agreement I received two errors - 'Page Limit Exceeded' (even though the PDF was only 1 page) and 'Document Upload Problem'. 

Adding in the ContentType header on the HttpContent of the PDF file fixed this issue.